% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{clamp}
\alias{clamp}
\title{clamp}
\usage{
clamp(x, lower = -Inf, upper = Inf)
}
\arguments{
\item{x}{Value or array of values to clamp}

\item{lower}{lower bound of the clamp}

\item{upper}{upper bound of the clamp}
}
\description{
Bound a value or array of values
}
\examples{
clamp(1:5) # returns c(1,2,3,4,5)
clamp(1:5, lower=3) # returns c(3,3,3,4,5)
}
