% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{DB_DE_regression_model}
\alias{DB_DE_regression_model}
\title{Create and return a GLM object from DB and DE Data}
\usage{
DB_DE_regression_model(merged_df)
}
\arguments{
\item{merged_df}{Merged data generated from running \code{merge_diffbind_with_DE()}.}
}
\value{
An object of class inheriting from 'glm'.
}
\description{
DB_DE_regression_model(merged_df) fits a generalized linear model (GLM) to test
the association between DE and DB data. It regresses \code{DE.logFC} against
\code{DB.logFC} in a merged data frame.
}
\details{
The summary of the regression is printed to the console.
}
\examples{
DB_DE_regression_model(merged_df = merged_df)
}
