% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_edger_analyzer.R
\name{draw_volcano}
\alias{draw_volcano}
\title{Draw a volcano plot}
\usage{
draw_volcano(lrt, figure_title = NULL, fdr = 0.01, xdiff = 5, ymax = 40)
}
\arguments{
\item{lrt}{A dataframe of the edgeR data to plot, typically generated by \code{perform_edger}}

\item{figure_title}{Title of the volcano plot(ie \verb{Smc3 cKO vs Control})}

\item{fdr}{The False Discovery Rate threshold at which to color genes}

\item{xdiff}{x-axis ceiling in both directions. For example \code{xdiff=5}
clamps out-of-bounds x points at -5 or 5}

\item{ymax}{y-axis ceiling. Floor is always 0. For example \code{ymax=40}
clamps out-of-bounds y points at 0 or 40}
}
\description{
draw_volcano() creates a volcano plot
from the data in \code{lrt} and returns the ggplot object but has no side effects.
It does NOT save the plot
}
\examples{
draw_volcano(smc3.lrt)
}
\keyword{plot}
\keyword{volcano}
