% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{draw_volcano_general}
\alias{draw_volcano_general}
\title{Create and return a volcano plot}
\usage{
draw_volcano_general(
  data,
  colors,
  size = 1,
  alpha = 1,
  criteria = NULL,
  xdiff = 5,
  ymax = 40
)
}
\arguments{
\item{data}{A \code{data.frame} containing the input data. Must include at least
the columns \code{Fold} (x-axis) and \code{negLogFDR} (y-axis).}

\item{colors}{Character. Either:
\itemize{
\item A single unnamed color string if \code{criteria} is \code{NULL} that is
used for all points.
\item A named character vector of colors, where names correspond to grouping
labels in the specified
\code{criteria} column. For example: \code{setNames(c("grey", "tomato"),
    c("NS", "Sig"))}.
}}

\item{size}{Numeric. Point size in the plot. Default is \code{1}.}

\item{alpha}{Numeric. Transparency level of the points. Ranges from 0 (fully
transparent) to 1 (fully opaque). Default is \code{1}.}

\item{criteria}{Optional. Character string. The name of a column in \code{data}
that defines groupings for coloring. If \code{NULL}, all points use the same color.}

\item{xdiff}{Numeric. Limits the x-axis range to \code{[-xdiff, xdiff]}. Default is \code{5}.}

\item{ymax}{Numeric. Limits the y-axis range to \code{[0, ymax]}. Default is \code{40}.}
}
\value{
A \code{ggplot} object representing the volcano plot.
}
\description{
draw_volcano_general(data, colors, size, alpha, criteria, xdiff, ymax) creates
and returns a volcano plot. This function allows flexible coloring based on a
user-defined grouping column in the dataset or applies a uniform color if no
grouping is specified.
}
\examples{
draw_volcano_general(data=data, colors = "grey")
}
