% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{get_DBsites}
\alias{get_DBsites}
\title{Retrieve Differentially Bind Sites from a DBA object}
\usage{
get_DBsites(
  dbaobjpath,
  FDR_THRESHOLD = 0.05,
  contrastnumber = 1,
  filepath = NULL
)
}
\arguments{
\item{dbaobjpath}{Character. File path to the DBA object (RDS file) created from \code{perform_diffbind()}}

\item{FDR_THRESHOLD}{Numeric. Significance cutoff for filtering DB sites. Default is \code{0.05}.}

\item{contrastnumber}{Integer. Index specifying which contrast to extract results from. Defaults to \code{1},
corresponding to the first contrast.}

\item{filepath}{Character. Output directory where results are written.}
}
\description{
get_DBsites(dbaobjpath, FDR_THRESHOLD, contrastnumber, filepath) Extracts differentially
bound (DB) regions from a DBA object and writes the results into BED-formatted files. The
function filters regions based on a user-defined FDR threshold and saves separate files for
all significant sites, upregulated, and downregulated sites.
}
\examples{
get_DBsites(dbaobjpath=dbaobjpath)
}
