% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{get_diffbind_sites}
\alias{get_diffbind_sites}
\title{Retrieve Differentially Bind Sites from a DBA object}
\usage{
get_diffbind_sites(
  dba_object,
  fdr_threshold = 0.05,
  contrast_number = 1,
  verbose = TRUE
)
}
\arguments{
\item{dba_object}{DBA object obtained after \code{perform_diffbind()}.}

\item{fdr_threshold}{Numeric. Significance cutoff for filtering DB sites.}

\item{contrast_number}{Integer. Index specifying which contrast to extract
results from.}

\item{verbose}{Boolean. If true, it prints information on number of DB sites.
Silent if false.}
}
\value{
A dataframe that contains filtered sites information.
}
\description{
get_diffbind_sites(dba_object, fdr_threshold, contrast_number) Extracts differentially
bound (DB) regions from a DBA object and filters regions based on a user-defined
FDR threshold, finally returns a dataframe that contains sites information.
}
\examples{
get_diffbind_sites(dba_object = dba.obj)
}
