% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{get_diffbind_volcano_data}
\alias{get_diffbind_volcano_data}
\title{Generate volcano-plot-ready data from DiffBind results}
\usage{
get_diffbind_volcano_data(
  dba_object,
  contrast_number = 1,
  fdr_threshold = 0.05,
  xdiff = 3,
  ymax = 20
)
}
\arguments{
\item{dba_object}{DBA object (RDS format) generated by \code{perform_diffbind()}.}

\item{contrast_number}{Integer. Index indicating which contrast to extract from the
DBA object.}

\item{fdr_threshold}{Numeric. FDR cutoff for significance.}

\item{xdiff}{Numeric. Limits the x-axis range to \code{[-xdiff, xdiff]}.}

\item{ymax}{Numeric. Limits the y-axis range to \code{[0, ymax]}.}
}
\value{
A dataframe that contains volcano-plottable data.
}
\description{
get_diffbind_volcano_data(dba_object, contrast_number, fdr_threshold, xdiff, ymax) makes
volcano-plottable table and applies filtering based on FDR threshold and clamps
values.
}
\examples{
get_diffbind_volcano_data(dba_object = dba.obj)
}
