% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_edger_analyzer.R
\name{get_gtf}
\alias{get_gtf}
\title{Get reference}
\usage{
get_gtf(gtf, verbose = FALSE)
}
\arguments{
\item{gtf}{Absolute path to the reference}

\item{verbose}{Set to TRUE to print number of genes in the reference}
}
\value{
A dataframe of gene annotations with 3 columns: the ensembl
\code{gene_id}, the \code{gene_type}, and \code{gene_name} (gene symbol)
}
\description{
get_gtf(gtf, verbose) returns the reference specified by
the path gtf.
}
\examples{
get_gtf(gtf='reference/gtf/gencode.vM14.primary_assembly.ERCC.annotation.gtf')
}
