% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{ggsave_vector_raster}
\alias{ggsave_vector_raster}
\title{Save a figure to both PNG and PDF}
\usage{
ggsave_vector_raster(
  filename,
  width = NA,
  height = NA,
  vector_device = "pdf",
  raster_device = "png",
  ...
)
}
\arguments{
\item{filename}{The base filename of the images you want to save. DO NOT
include an extension}

\item{width}{Figure width passed to ggsave. Default NA}

\item{height}{Figure height passed to ggsave. Default NA}

\item{vector_device}{Device to use for the vector image. See ggsave
documentation for options. Default 'pdf'}

\item{raster_device}{Device to use for the raster image. See ggsave
documentation for options. Default 'png'}

\item{...}{Additional options passed to ggsave}
}
\value{
None
}
\description{
ggsave_vector_raster(filename, ...) is a wrapper to ggplot2::ggsave. Provide
a base file name and this will
save the most recently displayed figure to both a raster (default PNG) and
vector (default PDF) format in the same location
}
\examples{
saved_plot = ggplot2::ggplot(mtcars, aes(cyl, mpg)) + ggplot2::geom_point()
saved_plot

ggsave_vector_raster('figure_1A')
ggsave_vector_raster('figure_2A', width=8, height=6, vector_device='svg', raster_device='jpeg')

# Pass extra arguments to ggsave() through ...
ggsave_vector_raster('figure_1A', plot=saved_plot) 
}
