% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_edger_analyzer.R
\name{label_con}
\alias{label_con}
\title{Adds a condition column which labels samples as knockout or control}
\usage{
label_con(sampleTable, col, pattern)
}
\arguments{
\item{sampleTable}{table we want to annotate}

\item{col}{column to search}

\item{pattern}{string or regex of which samples are controls}
}
\value{
A copy of sampleTable with the new column \code{condition} where the value is either ko or CON
}
\description{
label_con(sampleTable, col, pattern) labels a sample as a control if the
value in col matches pattern
}
\examples{
label_con(smc3, "Condition", "ctrl")
}
