% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_edger_analyzer.R
\name{label_control_samples}
\alias{label_control_samples}
\title{Adds a condition column which labels samples as knockout or control}
\usage{
label_control_samples(sampleTable, col, pattern)
}
\arguments{
\item{sampleTable}{Table we want to annotate}

\item{col}{Column to search}

\item{pattern}{String or regex of which samples are controls}
}
\value{
A copy of sampleTable with the new column \code{condition} where the value
is either ko or CON
}
\description{
label_control_samples(sampleTable, col, pattern) labels a sample as a control if the
value in col matches pattern
}
\examples{
label_control_samples(smc3, "Condition", "ctrl")
}
