% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_edger_analyzer.R
\name{label_volcano_plot_genes}
\alias{label_volcano_plot_genes}
\title{Label genes on a volcano plot}
\usage{
label_volcano_plot_genes(p, lrt, label_genes)
}
\arguments{
\item{p}{A volcano plot ggplot object}

\item{lrt}{An edgeR object with Likelihood Ratio Tests performed}

\item{label_genes}{Vector of gene names to label}

\item{...}{Arguments to be passed to \code{ggrepel::geom_label_repel}}
}
\description{
label_volcano_plot_genes() labels the specified genes in the provided
volcano plot and returns the ggplot object
}
\examples{
label_volcano_plot_genes(smc3.volcano, smc3.edger, c('Cars','Cars2'))
}
\keyword{genes}
\keyword{label}
\keyword{volcano}
