% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{make_density_plot}
\alias{make_density_plot}
\title{Create and save a density plot}
\usage{
make_density_plot(
  dbaobjpath,
  figure_title,
  filename,
  figure_dir = NULL,
  contrastnumber = 1,
  FDR_THRESHOLD = 0.05,
  width = 8,
  height = 6,
  DENSITY_color = "aquamarine4"
)
}
\arguments{
\item{dbaobjpath}{Character. Path to the saved DBA object (RDS format) from \code{perform_diffbind()}.}

\item{figure_title}{Character. Title displayed at the top of the plot.}

\item{filename}{Character. Base name (without extension) used for saving the plot files.}

\item{figure_dir}{Character. Directory path where the plot files will be saved. Default
is set to \code{NULL}.}

\item{contrastnumber}{Integer. Index indicating which contrast to extract from the DBA
object. Defeault is set to \code{1}.}

\item{FDR_THRESHOLD}{Numeric. The FDR threshold used for annotation purposes in the plot
title. Default is \code{0.05}.}

\item{width}{Numeric. Width dimensions of saved plot in inches. Default is 8.}

\item{height}{Numeric. Height dimensions of saved plot in inches. Default is 6.}

\item{DENSITY_color}{Character. Color to fill the density plot. Default is \code{"aquamarine4"}.}
}
\value{
A \code{ggplot2} object containing the generated density plot.
}
\description{
make_density_plot(dbaobjpath, figure_title, filename, figure_dir, contrast_number,
FDR_THRESHOLD, width, height, DENSITY_color) creates a density plot from a DBA
object with specified figure title and saves it in pdf and png format under the
designated file path.
}
\examples{
make_density_plot(
  dbaobjpath=dbaobjpath,
  figure_title="Distribution of Differential Binding Regions",
  figure_dir="diffbind_figures"),
  filename="df_fold_density_h3k27ac")
}
\keyword{Density}
\keyword{Plot}
