% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{make_diffbind_density_plot}
\alias{make_diffbind_density_plot}
\title{Create and save a density plot}
\usage{
make_diffbind_density_plot(
  dba_object,
  figure_title,
  save_name,
  fdr_threshold = 1,
  save_directory = NULL,
  contrast_number = 1,
  xdiff = 3,
  width = 8,
  height = 6,
  color = "aquamarine4"
)
}
\arguments{
\item{dba_object}{DBA object obtained after \code{perform_diffbind()}.}

\item{figure_title}{Character. Title for density plot.}

\item{save_name}{Character. Base name used for saving the plot files.}

\item{fdr_threshold}{Numeric. FDR cutoff for DB significance.}

\item{save_directory}{Character. Directory path where the plot files will be saved.}

\item{contrast_number}{Integer. Index indicating which contrast to extract from the
DBA object.}

\item{xdiff}{Numeric. Limits the x-axis range to \code{[-xdiff, xdiff]}. Setting to
\code{NULL} frees the limits.}

\item{width}{Numeric. Width dimensions of saved plot in inches.}

\item{height}{Numeric. Height dimensions of saved plot in inches.}

\item{color}{Character. Color to fill the density plot.}
}
\value{
A ggplot object of density plot.
}
\description{
make_diffbind_density_plot(dba_object, figure_title, save_name, fdr_threshold, save_directory,
contrast_number, xdiff, width, height, color) creates a density plot from a DBA object
with specified figure title and saves it in pdf and png format under the designated
file path.
}
\examples{
make_diffbind_density_plot(
  dba_object = dba.obj,
  figure_title = "Distribution of Differential Binding Regions",
  save_name = "df_fold_density_h3k27ac",
  save_directory = "diffbind_figures")
}
\keyword{Density}
\keyword{Plot}
