% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{make_diffbind_volcano_plot}
\alias{make_diffbind_volcano_plot}
\title{Create and save a volcano plot}
\usage{
make_diffbind_volcano_plot(
  dba_object,
  figure_title,
  save_name,
  save_directory = NULL,
  point_size = 1,
  point_alpha = 1,
  width = 8,
  height = 6,
  xdiff = 3,
  ymax = 20,
  contrast_number = 1,
  fdr_threshold = 0.05,
  color = c("grey", "aquamarine4")
)
}
\arguments{
\item{dba_object}{DBA object (RDS format) generated by \code{perform_diffbind()}.}

\item{figure_title}{Character. Figure title for volcano plot, ex. "h3k27me3 cKO vs cHET"}

\item{save_name}{Character. Base name for the output volcano plot figures.}

\item{save_directory}{Character. Directory path where the output plots will be saved.}

\item{point_size}{Numeric. Size of the points in volcano plots.}

\item{point_alpha}{Numeric. Transparency level of the points. Ranges from 0 (fully
transparent) to 1 (fully opaque).}

\item{width}{Numeric. Width dimensions of saved plot in inches.}

\item{height}{Numeric. Height dimensions of saved plot in inches.}

\item{xdiff}{Numeric. Limits the x-axis range to \code{[-xdiff, xdiff]}.}

\item{ymax}{Numeric. Limits the y-axis range to \code{[0, ymax]}.}

\item{contrast_number}{Integer. Index indicating which contrast to extract from the DBA object.}

\item{fdr_threshold}{Numeric. The FDR threshold used for grouping.}

\item{color}{Character vector of length 2. One for insignificant DB sites, one for significant
DB sites.}
}
\value{
A ggplot object of volcano plot
}
\description{
make_diffbind_volcano_plot(dba_object, figure_title, save_name, save_directory,
point_size, point_alpha, width, height, xdiff, ymax, contrast_number, fdr_threshold, color)
generates volcano plot and returns the ggplot object and it also saves the file
under the designated figure file path.
}
\examples{
make_diffbind_volcano_plot(
  dba_object = dba.obj,
  figure_title = "h3k27ac cKO vs cHET",
  save_directory = "diffbind_figures",
  save_name = "db_volcano")
}
\keyword{plot}
\keyword{volcano}
