% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{make_diffbind_volcano_plot_from_merged}
\alias{make_diffbind_volcano_plot_from_merged}
\title{Make Volcano Plot from DB and DE Data (Color by DE logFC direction)}
\usage{
make_diffbind_volcano_plot_from_merged(
  merged_df,
  figure_title,
  save_directory,
  save_name,
  xdiff = 3,
  ymax = 20,
  point_size = 1,
  point_alpha = 1,
  width = 8,
  height = 6,
  DB_fdr_cutoff = 0.05,
  DE_fdr_cutoff = 0.05,
  color = c("steelblue", "tomato")
)
}
\arguments{
\item{merged_df}{Merged data generated from running \code{merge_diffbind_with_DE()}}

\item{figure_title}{Character. The title of the volcano plot.}

\item{save_directory}{Character. The path to save the volcano plot.}

\item{save_name}{Character. Output figure save_name (e.g., "volcano.png").}

\item{xdiff}{Numeric. Limits the x-axis range to \code{[-xdiff, xdiff]}.}

\item{ymax}{Numeric. Limits the y-axis range to \code{[0, ymax]}.}

\item{point_size}{Numeric. Point size in the plot. Default is \code{1}.}

\item{point_alpha}{Numeric. Transparency level of the points. Ranges from 0 (fully
transparent) to 1 (fully opaque). Default is \code{1}.}

\item{width}{Numeric. Width dimensions of saved plot in inches. Default is 8.}

\item{height}{Numeric. Height dimensions of saved plot in inches. Default is 6.}

\item{DB_fdr_cutoff}{Numeric. Significance threshold for FDR in differential binding
dataframe. Default is 0.05.}

\item{DE_fdr_cutoff}{Numeric. Significance threshold for FDR in RNA-seq differential
expression. Defualt is 0.05.}

\item{color}{Character vector of length 2, in order of down FC color, then up FC
color. default is set to \code{c('steelblue', 'tomato')}.}
}
\value{
A ggplot object of DB and DE merged volcano plot
}
\description{
make_diffbind_volcano_plot_from_merged(merged_df, figure_title, save_directory, save_name,
xdiff, ymax, point_size, point_alpha, width, height, DB_fdr_cutoff, DE_fdr_cutoff,
color) creates and saves a volcano plot based on merged DB and DE data. All sites
are shown in gray; DB sites mapped to significant DE genes are colored by direction
of DE logFC (Up/Down).
}
\examples{
make_diffbind_volcano_plot_from_merged(
  merged_df = merged_volcano,
  figure_title = 'cKO vs cHet Differential Binding sites',
  save_directory = "diffbind_figures",
  save_name = "db_volcano_DE_color_binary")
}
