% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_edger_analyzer.R
\name{make_MA_plot}
\alias{make_MA_plot}
\title{Create and save a ma plot}
\usage{
make_MA_plot(
  lrt,
  figure_title,
  filename,
  figure_dir,
  fdr = 0.01,
  xmax = NULL,
  ydiff = 5,
  smooth_line = FALSE
)
}
\arguments{
\item{lrt}{lrt A dataframe of the edgeR data to plot, typically generated by
\code{perform_edger}}

\item{figure_dir}{Directory to save the plots under}

\item{fdr}{False discovery rate threshold. Genes <= fdr are colored}

\item{xmax}{x-axis ceiling}

\item{ydiff}{y-axis ceiling in both directions. For example, ydiff=5 sets the
y axis to -5 to 5}

\item{smooth_line}{line of estimate of mean, FALSE shows no smooth line, TRUE
otherwise}

\item{ma_figure_title}{The title of the MA plot (ex. logCPM vs logFC)}

\item{file_name}{The base name for the saved plot file}
}
\description{
make_MA_plot(lrt, figure_title, filename, figure_dir, fdr, xmax, ydiff) creates a
MA plot from the data in lrt with title figure_title and saves it in
a pdf and png under the folder figure_dir
}
\examples{
make_MA_plot(lrt =lrt, figure_title="logCPM vs logFC", filename = "test_make_ma_plot",
figure_dir = "informal_tests")
}
\keyword{ma}
\keyword{plot}
