% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_edger_analyzer.R
\name{make_PCA_plot}
\alias{make_PCA_plot}
\title{Create and save a PCA plot}
\usage{
make_PCA_plot(
  lrt,
  color = NULL,
  with_legend = TRUE,
  figure_title = NULL,
  figure_dir,
  filename
)
}
\arguments{
\item{lrt}{lrt A dataframe of the edgeR data to plot, typically generated by
\code{perform_edger}}

\item{color}{A vector that contains colors of the choice, the length of
the vector should be equal to the number of unique elements in label_list. By
default(i.e. NULL), use ggplot2 default colors}

\item{with_legend}{Boolean values, if True, draw the legend on the PCA, if false,
no legend. Default is set to TRUE}

\item{figure_title}{The title of the PCA plot (ex. Principal Component Analysis)}

\item{figure_dir}{Directory to save the plots under}

\item{filename}{The base name for the saved plot file}
}
\description{
make_PCA_plot <- function(lrt, color, with_legend=TRUE, figure_title=NULL,
figure_dir, filename) creates a PCA plot from the data in lrt with title
figure_title and saves it in a pdf and png under the folder figure_dir
}
\examples{
make_PCA_plot(lrt, color = c("blue", "red"), figure_title = "Principal Component Analysis", 
figure_dir = "PCA", filename = "plot")
make_PCA_plot(lrt, with_legend = FALSE, figure_title = "Principal Component Analysis", 
figure_dir = "PCA", filename = "plot")
}
\keyword{PCA}
\keyword{plot}
