% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_edger_analyzer.R
\name{make_volcano}
\alias{make_volcano}
\title{Create a volcano plot}
\usage{
make_volcano(
  lrt,
  figure_title,
  filename,
  figure_dir,
  fdr = 0.01,
  xdiff = 5,
  ymax = 40,
  label_genes = NULL
)
}
\arguments{
\item{lrt}{a dataframe of the egdger data to plot, typically generated by \code{perform_edger}}

\item{figure_title}{name of the gene being plotted.}

\item{filename}{name the plot is saved under}

\item{figure_dir}{directory to save the plots under}

\item{fdr}{false discovery rate threshold. genes > fdr are colored Default is 0.01}

\item{xdiff}{x-axis ceiling in both directions, For example, xdiff=5 sets the x axis to -5 to 5. Default is 5}

\item{ymax}{y-axis ceiling. Default is 40}

\item{label_genes}{list of genes to label on the volcano plot.}
}
\description{
make_volcano(lrt, figure_title, filename, figure_dir, fdr, xdiff,
ymax, intersect, intersect_color, label_genes) creates a volcano plot from the
data in lrt with title volcano_figure_title and saves it in a pdf and png
under the folder figure_dir
}
\examples{
make_volcano(smc3, figure_title="Smc3", filename="edger_smc3",
figure_dir=figure_dir)
}
\keyword{plot}
\keyword{volcano}
