% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_edger_analyzer.R
\name{make_volcano_dataframe}
\alias{make_volcano_dataframe}
\title{Make the dataframe containing volcano-plottable data}
\usage{
make_volcano_dataframe(lrt, xdiff = 5, ymax = 40)
}
\arguments{
\item{lrt}{An edgeR object with Likelihood Ratio Tests performed}

\item{xdiff}{x-axis ceiling in both directions. For example \code{xdiff=5}
clamps out-of-bounds x points at -5 or 5}

\item{ymax}{y-axis ceiling. Floor is always 0. For example \code{ymax=40}
clamps out-of-bounds y points at 0 or 40}
}
\description{
make_volcano_dataframe() computes the table of logFC and FDR values
from the \code{lrt} object and processes values for plotting
}
\keyword{data}
\keyword{volcano}
