% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{make_volcano_plot_diffbind}
\alias{make_volcano_plot_diffbind}
\title{Create and save a volcano plot}
\usage{
make_volcano_plot_diffbind(
  volcanopath,
  sigsitespath,
  figure_title,
  figure_dir,
  filename,
  size = 1,
  alpha = 1,
  width = 8,
  height = 6,
  xdiff = 5,
  ymax = 40,
  FDR_THRESHOLD = 0.05,
  color = c("aquamarine4", "grey")
)
}
\arguments{
\item{volcanopath}{Character. Path to the volcano_sites.bed generated by \code{make_volcano_sites()}.}

\item{sigsitespath}{Character. Path to the sig_sites.bed generated by \code{get_DBsites()}.}

\item{figure_title}{Character. Figure title for volcano plot, ex. h3k27me3 cKO vs cHET}

\item{figure_dir}{Character. Directory path where the output plots will be saved.}

\item{filename}{Character. Base name for the output figures.}

\item{size}{Numeric. Size of the points, default is set to 1.}

\item{alpha}{Numeric. Transparency level of the points. Ranges from 0 (fully
transparent) to 1 (fully opaque). Default is \code{1}.}

\item{width}{Numeric. Width dimensions of saved plot in inches. Default is 8.}

\item{height}{Numeric. Height dimensions of saved plot in inches. Default is 6.}

\item{xdiff}{Numeric. Limits the x-axis range to \code{[-xdiff, xdiff]}. Default is \code{5}.}

\item{ymax}{Numeric. Limits the y-axis range to \code{[0, ymax]}. Default is \code{40}.}

\item{FDR_THRESHOLD}{Numeric. The FDR threshold used for grouping, default is \code{0.05}.}

\item{color}{Character vector of length 2. By default the color is \code{c('aquamarine4', 'grey')}}
}
\value{
ggplot object
}
\description{
make_volcano_plot_diffbind(volcanopath, sigsitespath, figure_title, figure_dir,
filename, size, alpha, width, height, xdiff, ymax, FDR_THRESHOLD, color) generates
volcano plot from data in volcano_sites bed files and returns the ggplot object
also saves the file under the designated figure file path.
}
\examples{
make_volcano_plot_diffbind(volcanopath=volcanopath,
                           sigsitespath=sigsitespath,
                           figure_title="h3k27ac cKO vs cHET",
                           figure_dir="diffbind_figures"),
                           filename="db_volcano")
}
