% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{make_volcano_sites}
\alias{make_volcano_sites}
\title{Generate volcano-plot-ready data from DiffBind results}
\usage{
make_volcano_sites(
  dba_object,
  contrast_number = 1,
  fdr_threshold = 0.05,
  xdiff = 5,
  ymax = 40
)
}
\arguments{
\item{dba_object}{DBA object (RDS format) generated by \code{perform_diffbind()}.}

\item{contrast_number}{Integer. Index indicating which contrast to extract from the
DBA object. Defeault is set to \code{1}.}

\item{fdr_threshold}{Numeric. FDR cutoff for significance. Default is set to \code{0.05}.}

\item{xdiff}{Numeric. Limits the x-axis range to \code{[-xdiff, xdiff]}. Default is \code{5}.}

\item{ymax}{Numeric. Limits the y-axis range to \code{[0, ymax]}. Default is \code{40}.}
}
\value{
A dataframe that contains volcano-plottable data.
}
\description{
make_volcano_sites(dba_object, contrast_number, fdr_threshold, xdiff, ymax) makes
volcano-plottable table and applies filtering based on FDR threshold and clamps
values.
}
\examples{
make_volcano_sites(dba_object = dba.obj)
}
