% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{merge_sites_with_exp}
\alias{merge_sites_with_exp}
\title{Merge Differential Binding and Expression Data}
\usage{
merge_sites_with_exp(
  dba_object,
  RNAfile_path,
  join_data_sites,
  save_path,
  contrast_number = 1,
  fdr_threshold = 0.05,
  bulk_fdr_cutoff = 0.05
)
}
\arguments{
\item{dba_object}{DBA object (RDS format) generated by \code{perform_diffbind()}.}

\item{RNAfile_path}{Character. Path to the CSV file containing bulk RNA-seq differential
expression results. This file must include the columns \code{gene_name}, \code{logFC},
and \code{FDR}.}

\item{join_data_sites}{dataframe (e.g., volcano_sites) to be merged with DE data.}

\item{save_path}{Character. Directory path for storing intermediate files, including the TSS
BED file, sorted BED file, and nearest gene lookup TSV.}

\item{contrast_number}{Integer. Index indicating which contrast to extract from the
DBA object. Defeault is set to \code{1}.}

\item{fdr_threshold}{Numeric. Significance threshold for FDR in differential binding dataframe.
Default is 0.05.}

\item{bulk_fdr_cutoff}{Numeric. Significance threshold for RNA-seq differential expression.
Defualt is 0.05.}
}
\value{
A data frame containing merged binding and expression data. Each row corresponds to
a binding site and includes:
\itemize{
\item Site genomic coordinates
\item Site ID
\item Nearest gene name
\item \code{bulk.logFC} and \code{bulk.FDR} values
}
}
\description{
merge_sites_with_exp(dba_object, RNAfile_path, join_data_sites, save_path, contrast_number,
fdr_threshold, bulk_fdr_cutoff) integrates differential binding sites with differential
expression data (from bulk RNA-seq) by joining them based on their nearest gene TSS and
returns a data frame where each row represents a differential binding site along with its
nearest gene's differential expression.
}
\examples{
merge_sites_with_exp(
  dba_object = dba.obj,
  RNAfile_path = "path_to_RNAseq_csv_file",
  join_data_sites = volcano.sites,
  save_path = "diffbind_files")
}
