% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{perform_diffbind}
\alias{perform_diffbind}
\title{Perform Diffbind Analysis}
\usage{
perform_diffbind(
  samplesheetpath,
  reference,
  minMembers = NULL,
  analysismethod = DBA_DESEQ2,
  datatype = DBA_DATA_FRAME,
  FDRthreshold = 0.05,
  normalization = DBA_NORM_LIB,
  filesuffix = NULL,
  filePath = NULL
)
}
\arguments{
\item{samplesheetpath}{Character. The path to the sample CSV file.
Normally the file is named 'diffbind_samples.csv' and placed in the current
working directory. It must contain the following columns:
\itemize{
\item \strong{"Sample_ID"}
\item \strong{"Tissue"}
\item \strong{"Condition"}
\item \strong{"bamReads"}
\item \strong{"Peaks"}
\item \strong{"PeakCaller"}
}}

\item{reference}{Character. Reference group name (e.g., "HET", "Ctrl").}

\item{minMembers}{Integer. Minimum number of replicates per condition (default is \code{NULL}).
Set to 2 if any condition has only 2 replicates, otherwise set to \code{NULL}.}

\item{analysismethod}{The analysis method, either \code{DBA_DESEQ2} (default) or \code{DBA_EDGER}.}

\item{datatype}{Output format of peak data (default is \code{DBA_DATA_FRAME}).
Other options are \code{DBA_DATA_GRANGES}, \code{DBA_DATA_RANGEDDATA}.}

\item{FDRthreshold}{Numeric. FDR cutoff for significance, used in both filtering
and visualization (Default is 0.05).}

\item{normalization}{Normalization method (default is \code{DBA_NORM_LIB}). Other options
include \code{DBA_NORM_RLE}, \code{DBA_NORM_TMM}, etc.}

\item{filesuffix}{Character. Suffix for the output DBA object file.}

\item{filePath}{Character. Directory path to save the resulting DBA object (default is \code{NULL}).
If \code{NULL}, current working directory is used.}
}
\description{
perform_diffbind(samplesheetpath, reference, minMembers, analysismethod, datatype,
FDRthreshold, normalization, filesuffix, filePath) performs a complete DiffBind
analysis workflow using a provided sample sheet. It includes blacklist filtering,
counting, normalization, and contrast setup. The resulting DBA object is saved as
an RDS file under specific file path.
}
\examples{
perform_diffbind(samplesheetpath = "diffbind_samples.csv",
                 reference = "HET",
                 minMembers = 2,
                 filesuffix = "h3k27me3",
                 filepath = "diffbind_files")
}
