% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_edger_analyzer.R
\name{perform_edger}
\alias{perform_edger}
\title{Perform edgeR}
\usage{
perform_edger(
  sampleTable,
  fileCol = "Filename",
  idCol = "Sample",
  condCol = "condition",
  batchCol = NULL,
  filePrefix = NULL,
  gtf = NULL,
  saveName = NULL
)
}
\arguments{
\item{sampleTable}{A dataframe with sample data}

\item{fileCol}{Column name that refers to file names}

\item{idCol}{Column name that refers to the unique identifier of each sample}

\item{condCol}{Column name that refers to the experimental condition of each sample}

\item{batchCol}{(optional) To enable batch correction, pass a column name that
refers to the batch of the sample. If NULL, no batch correction}

\item{filePrefix}{Prefix to add to filenames (e.g. if they are in a different
directory). If NULL, search the current working directory}

\item{gtf}{A dataframe of the reference genome, generated by \code{kwanlibr::get_gtf}}

\item{saveName}{Filename to save edgeR tables under. Saves top 5k genes and
all genes. If NULL, do not save tables.}
}
\value{
A dataframe of the edgeR analysis
}
\description{
perform_edger(
sampleTable, fileCol, idCol, condCol, batchCol, filePrefix, gtf, saveName)
performs edgeR analysis on a dataframe of samples. Typically run after
get_gtf, subset_samples, and label_control_samples
}
\examples{
perform_edger(smc3, idCol="Pool.Name", gtf=gtf, saveName="tables/p21_smc3_edger")
}
