% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{save_diffbind_object}
\alias{save_diffbind_object}
\title{Save DBA object}
\usage{
save_diffbind_object(dba_object, file_suffix, save_directory = NULL)
}
\arguments{
\item{dba_object}{DBA object obtained after \code{perform_diffbind()}.}

\item{file_suffix}{Character. Suffix for the output DBA object file.}

\item{save_directory}{Character. Directory path to save the resulting
DBA object (default is \code{NULL}). If \code{NULL}, current working
directory is used.}
}
\description{
save_diffbind_object(dba_object, file_suffix, save_directory) saves the
DBA object under the specified file path. It returns nothing
}
\examples{
save_diffbind_object(
  dba_object = dba.obj,
  file_suffix = "h3k27me3",
  save_directory = "test_results")
}
