% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{save_diffbind_sites}
\alias{save_diffbind_sites}
\title{save differential binding sites files}
\usage{
save_diffbind_sites(
  dba_object,
  save_directory = NULL,
  fdr_threshold = 0.05,
  contrast_number = 1
)
}
\arguments{
\item{dba_object}{DBA object obtained after \code{perform_diffbind()}.}

\item{save_directory}{Character. Output directory where results are written.}

\item{fdr_threshold}{Numeric. FDR cutoff for significantly differntially
bound up and down sites.}

\item{contrast_number}{Integer. Index specifying which contrast to extract
results from.}
}
\description{
save_diffbind_sites(dba_object, save_directory, fdr_threshold, contrast_number) saves site
dataframes in bed format based on different fdr_threshold, which includes allDB
sites, upDB sites, and downDB sites. It doesn't return anything
}
\examples{
save_diffbind_sites(
  dba_object = dba.obj,
  save_directory = "diffbind_files")
}
