% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{save_sites}
\alias{save_sites}
\title{save different sites files
It doesn't return anything
save_sites(dba_object, save_path, fdr_threshold, contrast_number) saves site
dataframes in bed format based on different fdr_threshold, which includes allDB
sites, upDB sites, and downDB sites.}
\usage{
save_sites(dba_object, save_path, fdr_threshold = 0.05, contrast_number = 1)
}
\arguments{
\item{dba_object}{DBA object obtained after \code{perform_diffbind()}.}

\item{save_path}{Character. Output directory where results are written.}

\item{fdr_threshold}{Numeric. If it is 1, then it saves all DB sites, if
it is 0.05, then it save upDB and downDB sites separately.}

\item{contrast_number}{Integer. Index specifying which contrast to extract
results from. Defaults to \code{1} corresponding to the first contrast.}
}
\description{
save different sites files
It doesn't return anything
save_sites(dba_object, save_path, fdr_threshold, contrast_number) saves site
dataframes in bed format based on different fdr_threshold, which includes allDB
sites, upDB sites, and downDB sites.
}
\examples{
save_sites(
  dba_object = dba.obj,
  save_path = "diffbind_files")
}
