% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_edger_analyzer.R
\name{subset_table}
\alias{subset_table}
\title{Get a subset of table rows}
\usage{
subset_table(sampleTable, col, ...)
}
\arguments{
\item{sampleTable}{table we want to subset}

\item{col}{column to search}

\item{...}{All other arguments are strings or regex that are selected for in
the column. The result is the union of these searches}
}
\value{
A dataframe with the subset of selected rows
}
\description{
subset_table(sampleTable, col, ...) selects a subset of rows from
sampleTable where the value of col is in ...
}
\examples{
subset_table(samples, "Pool.Name", "Smc", "p21") # Selects only rows with "Smc" or "p21" in the Pool.name column
}
