% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{draw_PCA}
\alias{draw_PCA}
\title{draw_PCA()}
\usage{
draw_PCA(df, label_list = NULL, color_list = NULL, legend = TRUE)
}
\arguments{
\item{df}{A data-matrix or data-frame}

\item{label_list}{A vector that contains assigned groups of the sample, the
length of the vector should be the same as the number of samples in the df}

\item{color_list}{A vector that contains colors of the choice, the length of
the vector should be equal to the number of unique elements in label_list. By
default, use ggplot default colors}

\item{legend}{Boolean values, if True, draw the legend on the PCA, if false,
no legend. Default is set to TRUE}
}
\description{
run PCA and draw PCA plot
}
\examples{
draw_PCA(iris[-5], label_list = iris$Species, color_list = c("green", "blue", "yellow"))
draw_PCA(iris[-5], label_list = iris$Species)
draw_PCA(iris[-5], legend=FALSE)
draw_PCA(iris[-5]) if no label_list, no legend is produced
}
