% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{get_DBsites}
\alias{get_DBsites}
\title{Retrieve Differentially Bind Sites from a DBA object}
\usage{
get_DBsites(dba_object, fdr_threshold = 0.05, contrast_number = 1)
}
\arguments{
\item{dba_object}{DBA object obtained after \code{perform_diffbind()}.}

\item{fdr_threshold}{Numeric. Significance cutoff for filtering DB sites.
Default is \code{0.05}.}

\item{contrast_number}{Integer. Index specifying which contrast to extract
results from. Defaults to \code{1} corresponding to the first contrast.}
}
\value{
A dataframe that contains filtered sites information.
}
\description{
get_DBsites(dba_object, fdr_threshold, contrast_number) Extracts differentially
bound (DB) regions from a DBA object and filters regions based on a user-defined
FDR threshold, finally returns a dataframe that contains sites information.
}
\examples{
get_DBsites(dba_object = dba.obj)
}
