% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_edger_analyzer.R
\name{make_PCA_plot}
\alias{make_PCA_plot}
\title{Create and save a PCA plot}
\usage{
make_PCA_plot(
  lrt,
  filename,
  figure_dir,
  color = NULL,
  label = FALSE,
  legend = TRUE,
  figure_title = NULL
)
}
\arguments{
\item{lrt}{An edgeR object with Likelihood Ratio Tests performed, typically generated by
\code{perform_edger}}

\item{filename}{The base name for the saved plot file}

\item{figure_dir}{Directory to save the plots under}

\item{color}{A vector that contains colors of the choice, the length of
the vector should reference the number of unique conditions in the LRT object. By
default(i.e. NULL), use ggplot2 default colors}

\item{label}{Boolean value, if True, label each individual point by
individual labels, false shows no label. Default is set to FALSE}

\item{legend}{Boolean value, if True, draw the legend on the PCA, if false,
no legend. Default is set to TRUE}

\item{figure_title}{The title of the PCA plot (ex. Principal Component Analysis)}
}
\description{
make_PCA_plot <- function(lrt, color, with_legend=TRUE, figure_title=NULL,
figure_dir, filename) creates a PCA plot from the data in lrt with title
figure_title and saves it in a pdf and png under the folder figure_dir
}
\examples{
make_PCA_plot(lrt, color = c("blue", "red"), figure_title = "Principal Component Analysis", 
figure_dir = "PCA", filename = "plot")
make_PCA_plot(lrt, legend = FALSE, figure_title = "Principal Component Analysis", 
figure_dir = "PCA", filename = "plot")
}
\keyword{PCA}
\keyword{plot}
