% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{make_scatter_plot_from_merged}
\alias{make_scatter_plot_from_merged}
\title{Create and save the scattor plot from DB and DE Data}
\usage{
make_scatter_plot_from_merged(
  allDBsitepath,
  RNAfilepath,
  sigsitespath,
  filepath,
  figure_title,
  figure_path,
  filename,
  width = 8,
  height = 6,
  size = 1,
  regression = TRUE,
  BULK_FDR_CUTOFF = 0.05,
  point_color = "black",
  line_color = "aquamarine4"
)
}
\arguments{
\item{allDBsitepath}{Character. Path to BED file of all differential binding sites.}

\item{RNAfilepath}{Character. Path to RNA-seq results CSV (must include gene_name, logFC, FDR).}

\item{sigsitespath}{Character. Path to BED file of significant sites.}

\item{filepath}{Character. Directory path for storing intermediate files, including
the TSS BED file, sorted BED file, and nearest gene lookup TSV.}

\item{figure_title}{Character. The scatter plot title.}

\item{figure_path}{Character. Directory to save the figure output.}

\item{filename}{Character. The name of the file.}

\item{width}{Numeric. Width dimensions of saved plot in inches. Default is 8.}

\item{height}{Numeric. Height dimensions of saved plot in inches. Default is 6.}

\item{size}{Numeric. The size of the point in scatter plot. Default is 1.}

\item{regression}{logical value. Default is set to 'TRUE', a regression line is
drawn to the scatter plot. otherwise no regression line.}

\item{BULK_FDR_CUTOFF}{Numeric FDR threshold to filter DE genes. Default is 0.05.}

\item{point_color}{Character Value. The color of the point. Default is \code{'black'}.}

\item{line_color}{Character Value. The color of the regression line. Default is
\code{'aquamarine4'}.}
}
\description{
make_scatter_plot_from_merged() create scatter plot from merged differential
binding sites with differential expression data and save the plot under designated
directory.
}
\examples{
make_scatter_plot_from_merged()
}
