% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{make_scatter_plot_from_merged}
\alias{make_scatter_plot_from_merged}
\title{Create and save the scattor plot from DB and DE Data}
\usage{
make_scatter_plot_from_merged(
  merged_df,
  figure_title,
  save_directory,
  save_name,
  width = 8,
  height = 6,
  point_size = 1,
  regression = TRUE,
  DB_fdr_cutoff = 0.05,
  DE_fdr_cutoff = 1,
  point_color = "black",
  line_color = "aquamarine4"
)
}
\arguments{
\item{merged_df}{Merged data generated from running \code{merge_diffbind_with_DE()}}

\item{figure_title}{Character. The scatter plot title.}

\item{save_directory}{Character. Directory to save the figure output.}

\item{save_name}{Character. The base name of the scatter plot file.}

\item{width}{Numeric. Width dimensions of saved plot in inches. Default is 8.}

\item{height}{Numeric. Height dimensions of saved plot in inches. Default is 6.}

\item{point_size}{Numeric. The size of the point in scatter plot. Default is 1.}

\item{regression}{logical value. Default is set to 'TRUE', a regression line is
drawn to the scatter plot. otherwise no regression line.}

\item{DB_fdr_cutoff}{Numeric. Significance threshold for FDR in differential binding
dataframe. Default is 0.05.}

\item{DE_fdr_cutoff}{Numeric. Significance threshold for FDR in RNA-seq differential
expression. Default is 1.}

\item{point_color}{Character Value. The color of the point. Default is \code{'black'}.}

\item{line_color}{Character Value. The color of the regression line. Default is
\code{'aquamarine4'}.}
}
\value{
ggplot object of scatter plot
}
\description{
make_scatter_plot_from_merged(merged_df, figure_title, save_directory, save_name,
width, height, point_size, regression, DB_fdr_cutoff, DE_fdr_cutoff, point_color,
line_color) create and saves scatter plot from merged DB and DE data and save the
plot under designated directory.
}
\examples{
make_scatter_plot_from_merged(
  merged_df = merged_df,
  figure_title = "DB vs DE log2FC",
  save_directory = "diffbind_figures",
  save_name = "db_DE_scatter")
}
