% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_edger_analyzer.R
\name{make_volcano}
\alias{make_volcano}
\title{Create and save a volcano plot}
\usage{
make_volcano(
  lrt,
  figure_title,
  filename,
  figure_dir,
  fdr = 0.01,
  xdiff = 5,
  ymax = 40,
  label_genes = NULL
)
}
\arguments{
\item{lrt}{A dataframe of the edgeR data to plot, typically generated by \code{perform_edger}}

\item{figure_title}{Title of the volcano plot (ie \verb{Smc3 cKO})}

\item{filename}{Name the plot is saved under}

\item{figure_dir}{Directory to save the plots under}

\item{fdr}{False discovery rate threshold. Genes > fdr are colored}

\item{xdiff}{x-axis ceiling in both directions. For example, xdiff=5 sets the
x axis to -5 to 5}

\item{ymax}{y-axis ceiling}

\item{label_genes}{List of genes to label on the volcano plot}
}
\description{
make_volcano(lrt, figure_title, filename, figure_dir, fdr, xdiff,
ymax, intersect, intersect_color, label_genes) creates a volcano plot from the
data in lrt with title volcano_figure_title and saves it in a pdf and png
under the folder figure_dir
}
\examples{
make_volcano(smc3, figure_title="Smc3 cKO", filename="edger_smc3",
figure_dir=figure_dir)
}
\keyword{plot}
\keyword{volcano}
