% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{make_volcano_plot_diffbind}
\alias{make_volcano_plot_diffbind}
\title{Create and save a volcano plot}
\usage{
make_volcano_plot_diffbind(
  dba_object,
  figure_title,
  file_name,
  figure_save_path = NULL,
  point_size = 1,
  point_alpha = 1,
  width = 8,
  height = 6,
  xdiff = 5,
  ymax = 40,
  contrast_number = 1,
  fdr_threshold = 0.05,
  color = c("aquamarine4", "grey")
)
}
\arguments{
\item{dba_object}{DBA object (RDS format) generated by \code{perform_diffbind()}.}

\item{figure_title}{Character. Figure title for volcano plot, ex. "h3k27me3 cKO vs cHET"}

\item{file_name}{Character. Base name for the output volcano plot figures.}

\item{figure_save_path}{Character. Directory path where the output plots will
be saved. Default is \code{NULL}.}

\item{point_size}{Numeric. Size of the points in volcano plots, default is set to 1.}

\item{point_alpha}{Numeric. Transparency level of the points. Ranges from 0 (fully
transparent) to 1 (fully opaque). Default is \code{1}.}

\item{width}{Numeric. Width dimensions of saved plot in inches. Default is 8.}

\item{height}{Numeric. Height dimensions of saved plot in inches. Default is 6.}

\item{xdiff}{Numeric. Limits the x-axis range to \code{[-xdiff, xdiff]}. Default is \code{5}.}

\item{ymax}{Numeric. Limits the y-axis range to \code{[0, ymax]}. Default is \code{40}.}

\item{contrast_number}{Integer. Index indicating which contrast to extract from
the DBA object. Defeault is set to \code{1}.}

\item{fdr_threshold}{Numeric. The FDR threshold used for grouping, default is \code{0.05}.}

\item{color}{Character vector of length 2. One for insignificant DB sites, one for significant
DB sites. Usually 'grey' is for insignificant sites. By default, it is \code{c('aquamarine4', 'grey')}}
}
\value{
A ggplot object of volcano plot
}
\description{
make_volcano_plot_diffbind(dba_object, figure_title, file_name, figure_save_path,
point_size, point_alpha, width, height, xdiff, ymax, contrast_number, fdr_threshold, color)
generates volcano plot and returns the ggplot object and it also saves the file
under the designated figure file path.
}
\examples{
make_volcano_plot_diffbind(
  dba_object = dba.obj,
  figure_title = "h3k27ac cKO vs cHET",
  figure_save_path = "diffbind_figures",
  file_name = "db_volcano")
}
\keyword{plot}
\keyword{volcano}
