% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{make_volcano_plot_from_merged}
\alias{make_volcano_plot_from_merged}
\title{Make Volcano Plot from DB and DE Data (Color by bulk logFC direction)}
\usage{
make_volcano_plot_from_merged(
  allDBsitepath,
  RNAfilepath,
  volcanopath,
  filepath,
  figure_title,
  figure_path,
  filename,
  color,
  xdiff = 5,
  ymax = 40,
  size = 1,
  alpha = 1,
  width = 8,
  height = 6,
  BULK_FDR_CUTOFF = 0.05
)
}
\arguments{
\item{allDBsitepath}{Character. Path to BED file with differential binding sites.}

\item{RNAfilepath}{Character. Path to RNA-seq results CSV (must include gene_name, logFC, FDR).}

\item{volcanopath}{Character. Path to BED-like file of all sites (for background).}

\item{filepath}{Character. Directory to save intermediate and output files.}

\item{figure_title}{Character. The title of the volcano plot.}

\item{filename}{Character. Output figure filename (e.g., "volcano.png").}

\item{color}{Character vector of length 3, in order of background color, down FC color,
and up FC color. E.x. \code{c('grey','steelblue', 'tomato')}.}

\item{xdiff}{Numeric. Limits the x-axis range to \code{[-xdiff, xdiff]}. Default is \code{5}.}

\item{ymax}{Numeric. Limits the y-axis range to \code{[0, ymax]}. Default is \code{40}.}

\item{size}{Numeric. Point size in the plot. Default is \code{1}.}

\item{alpha}{Numeric. Transparency level of the points. Ranges from 0 (fully
transparent) to 1 (fully opaque). Default is \code{1}.}

\item{width}{Numeric. Width dimensions of saved plot in inches. Default is 8.}

\item{height}{Numeric. Height dimensions of saved plot in inches. Default is 6.}

\item{BULK_FDR_CUTOFF}{Numeric FDR threshold to filter DE genes. Default is set
to 0.05.}
}
\value{
A ggplot object.
}
\description{
make_volcano_plot_from_merged(allDBsitepath, RNAfilepath, volcanopath, filepath,
figure_title, filename, color, xdiff, ymax, size, alpha, width, height, BULK_FDR_CUTOFF)
merges differential binding (DB) sites with RNA-seq differential expression (DE)
data using \code{merge_sites_with_exp()}, then creates a volcano plot. All sites
are shown in gray; DB sites mapped to significant DE genes are colored by direction
of bulk logFC (Up/Down). Plot is saved and returned.
}
\examples{
make_volcano_plot_from_merged(allDBsitepath = "diffbind_files/allDB_sites.bed",
                              RNAfilepath="",
                              volcanopath="diffbind_files/volcano_sites.bed",
                              filepath="diffbind_files",
                              figure_title='cKO vs cHet Differential Binding sites',
                              figure_path="diffbind_figures",
                              filename="db_volcano_bulk_color_binary",
                              color=c('grey','steelblue', 'tomato'))
}
