% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{make_volcano_sites}
\alias{make_volcano_sites}
\title{Generate volcano-plot-ready data from DiffBind results}
\usage{
make_volcano_sites(
  dbaobjpath,
  sigsitespath,
  filepath,
  FDR_THRESHOLD = 0.05,
  xdiff = 5,
  ymax = 40
)
}
\arguments{
\item{dbaobjpath}{Character. Path to the saved DBA object (RDS format) generated by
\code{perform_diffbind()}.}

\item{sigsitespath}{Character. Path to the sig_sites.bed generated by \code{get_DBsites()}.}

\item{filepath}{Character. Output directory where volcano data is written.}

\item{FDR_THRESHOLD}{Numeric. FDR cutoff for significance. Default is set to 0.05.}

\item{xdiff}{Numeric. Limits the x-axis range to \code{[-xdiff, xdiff]}. Default is \code{5}.}

\item{ymax}{Numeric. Limits the y-axis range to \code{[0, ymax]}. Default is \code{40}.}
}
\description{
make_volcano_sites(dbaobjpath, sigsitespath, filepath, FDR_THRESHOLD, xdiff, ymax)
makes volcano-plottable table and applies filtering based on FDR threshold and clamps
values.
It does not return the data.
}
\examples{
make_volcano_sites(dbaobjpath=dbaobjpath,
                   sigsitespath=sigsitespath,
                   filepath="diffbind_files")
}
