% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{merge_sites_with_exp}
\alias{merge_sites_with_exp}
\title{Merge Differential Binding and Expression Data}
\usage{
merge_sites_with_exp(
  allDBsitepath,
  RNAfilepath,
  joindatapath,
  filepath,
  BULK_FDR_CUTOFF = 0.05
)
}
\arguments{
\item{allDBsitepath}{Character. Path to the BED file containing all differentially
bound sites. This file is typically generated by \code{get_DBsites()}.}

\item{RNAfilepath}{Character. Path to the CSV file containing bulk RNA-seq differential
expression results. This file must include the columns \code{gene_name}, \code{logFC},
and \code{FDR}.}

\item{joindatapath}{Character. Path to the BED-like table (e.g., volcano_sites.bed) to
be annotated with expression data. This is the backbone data for visualization.}

\item{filepath}{Character. Directory path for storing intermediate files, including the TSS
BED file, sorted BED file, and nearest gene lookup TSV.}

\item{BULK_FDR_CUTOFF}{Numeric. Significance threshold for RNA-seq differential expression
(e.g., 0.05). Defualt is set as 0.05.}
}
\value{
A data frame containing merged binding and expression data. Each row corresponds to
a binding site and includes:
\itemize{
\item Site genomic coordinates
\item Site ID
\item Nearest gene name
\item \code{bulk.logFC} and \code{bulk.FDR} values
}
}
\description{
merge_sites_with_exp(allDBsitepath, RNAfilepath, joindatapath, filepath, BULK_FDR_CUTOFF)
integrates differential binding sites with differential expression data (from bulk RNA-seq)
by joining them based on their nearest gene TSS and returns a data frame where each row
represents a differential binding site along with its nearest gene's differential expression.
}
\examples{
merge_sites_with_exp(allDBsitepath = "diffbind_files/allDB_sites.bed",
                     volcanositepath = "diffbind_files/volcano_sites.bed",
                     RNAfilepath = "path_to_RNAseq_csv_file",
                     filepath = "diffbind_files")
}
