% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_diffbind_analyzer.R
\name{perform_diffbind}
\alias{perform_diffbind}
\title{Perform Diffbind Analysis}
\usage{
perform_diffbind(
  sample_sheet_path,
  control_level,
  min_members = NULL,
  analysis_method = DBA_DESEQ2,
  data_type = DBA_DATA_FRAME,
  fdr_threshold = 0.05,
  normalization = DBA_NORM_LIB
)
}
\arguments{
\item{sample_sheet_path}{Character. The path to the sample CSV file.
Normally the file is named 'diffbind_samples.csv' and placed in the current
working directory. It must contain the following columns:
\itemize{
\item \strong{"Sample_ID"}
\item \strong{"Tissue"}
\item \strong{"Condition"}
\item \strong{"bamReads"}
\item \strong{"Peaks"}
\item \strong{"PeakCaller"}
}}

\item{control_level}{Character. Reference group name (e.g., "HET", "Ctrl").}

\item{min_members}{Integer. Minimum number of replicates per condition (default is \code{NULL}).
Set to 2 if any condition has only 2 replicates, otherwise set to \code{NULL}.}

\item{analysis_method}{The analysis method, either \code{DBA_DESEQ2} (default) or \code{DBA_EDGER}.}

\item{data_type}{Output format of peak data (default is \code{DBA_DATA_FRAME}).
Other options are \code{DBA_DATA_GRANGES}, \code{DBA_DATA_RANGEDDATA}.}

\item{fdr_threshold}{Numeric. FDR cutoff for significance, used in both filtering
and visualization (Default is 0.05).}

\item{normalization}{Normalization method (default is \code{DBA_NORM_LIB}). Other options
include \code{DBA_NORM_RLE}, \code{DBA_NORM_TMM}, etc.}
}
\value{
analyzed DBA object
}
\description{
perform_diffbind(sample_sheet_path, control_level, min_members, analysis_method,
data_type, fdr_threshold, normalization) performs a complete DiffBind analysis
workflow using a provided sample sheet, which includes blacklist filtering, counting,
normalization, and contrast setup.
}
\examples{
perform_diffbind(
  sample_sheet_path = "diffbind_samples.csv",
  control_level = "HET",
  min_members = 2)
}
\keyword{diffbind}
